Add CircleCI config file and Dockerfile
The Dockerfile that has been used in the image kshitijkabeer/ros-sim
has been included. Also the config.yml file required for CircleCI to
work has also been included. This file contains 2 preliminary jobs, one
for preparing the ros workspace with all the required packages, and the
other for building this workspace. These are just alpha versions, they
might be riddled with bugs.