Add different models for different LED lights
Visual plugins can only be included in the <visual> tag of the model. It
is impossible to include a visual plugin in the <include> tag. Neither
is it possible to pass any parameters in the include tag, which can be
used to differentiate the color of the 2 LEDs. Hence new models had to
be made, each for different colors of the LEDs.